# Conflicts:
#	src/main/java/com/svydovets/bibirnate/demo/DemoApp.java
#	src/main/java/com/svydovets/bibirnate/session/impl/JdbcEntityDaoFactory.java
#	src/main/java/com/svydovets/bibirnate/session/impl/SessionImpl.java
#	src/test/java/com/svydovets/bibirnate/session/impl/BaseJdbcEntityDaoTest.java
#	src/test/java/com/svydovets/bibirnate/session/impl/SessionImplTest.java

+++ src/test/java/com/svydovets/bibirnate/entity/BoboEntity.java
@@ -10,12 +10,12 @@ import com.svydovets.bibirnate.annotation.Id;
 @Cacheable
 public class BoboEntity {
     @Id
-    private long id;
+    private int id;
 
     public BoboEntity() {
     }
 
-    public BoboEntity(long id) {
+    public BoboEntity(int id) {
         this.id = id;
     }
